Reshuffle some more
